move BeanRoleMatcher to the feature specific package
