+++ spring-lens-insight/src/main/java/com/sdlcpro/springlens/insight/bean/InfraBeanRoleMatcher.java
@@ -1,6 +1,5 @@
 package com.sdlcpro.springlens.insight.bean;
 
-import com.sdlcpro.springlens.insight.support.matcher.BeanRoleMatcher;
 import com.sdlcpro.springlens.model.bean.BeanRole;
 
 import java.util.EnumSet;
